+++ Porovnu/Common/Views/AmountConvertTextField.swift
@@ -10,19 +10,20 @@ import SwiftUI
 struct AmountConvertTextField: View {
 
     // MARK: - Private properties
+    
     @FocusState private var isFocusedState: Bool
     @State private var displayAmount: String = String()
 
     // MARK: - Public properties
 
-    @Binding var amount: Double
+    @Binding var amount: Int
     @State var isFocused = false
     let placeholder: String
     let type: TextFieldTypeEnum
     private var onFocusChange: ((Bool) -> Void)?
 
     init(
-        amount: Binding<Double>,
+        amount: Binding<Int>,
         placeholder: String = "0",
         type: TextFieldTypeEnum = .casual
     ) {
@@ -61,7 +62,7 @@ private extension AmountConvertTextField {
     func applyAmount(_ text: String) {
         if text.isEmpty {
             amount = 0
-        } else if let value = Double.amountFrom(text) {
+        } else if let value = Int.amountFrom(text) {
             amount = value
         }
     }
